Better draw timing to reduce flicker/tear
The problem of draw timing is not unique to st. All terminals have to
deal with it, and a new suggested standard tries to solve it. It's
called "Synchronized Updates" and it allows the application to tell the
terminal when the output "batch" is complete so that the terminal knows
not to draw partial output - hence "application sync".

The suggestion - by iTerm2 author - is available here:
https://gitlab.com/gnachman/iterm2/-/wikis/synchronized-updates-spec

This patch adds synchronized-updates/application-sync support in st. It
requires the auto-sync patch above installed first. This patch has no
effect except when an application uses the synchronized-update escape
sequences.

Note that currently there are very few terminals or applications which
support it, but one application which does support it is tmux since
2020-04-18. With this patch nearly all cursor flicker is eliminated in
tmux, and tmux detects it automatically via terminfo and enables it when
st is installed correctly.

see: https://st.suckless.org/patches/sync/